LayoutTests/imported/w3c/web-platform-tests/svg/animations/syncbase-remove-add-while-running-expected.txt


FAIL Remove/Add syncbase while animation is running assert_equals: Sync base triggered expected 100 but got 0
